pyproject.toml
[project]
name = "salah-reminder"
version = "0.1.0"
description = "Smart Salah reminder with Quran, Hadith, scoring, and AI chat"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue,webhooks]>=21.0,<22.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "asyncpg>=0.29,<1.0",
    "psycopg2-binary>=2.9,<3.0",
    "praytimes>=2.3,<3.0",
    "prayerpy>=0.1",
    "apscheduler>=3.10,<4.0",
    "anthropic>=0.80,<1.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pytz>=2024.1",
    "timezonefinder>=6.0,<7.0",
    "caldav>=2.2,<3.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
    "jinja2>=3.1,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
salah-reminder = "src.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
